rename namespace Clarkson_Core\Object to Clarkson_Core\WordPress_Object
