+++ src/WordPress_Object/Clarkson_User.php
@@ -3,7 +3,7 @@
  * Clarkson User.
  */
 
-namespace Clarkson_Core\Object;
+namespace Clarkson_Core\WordPress_Object;
 
 use Clarkson_Core\Objects;
