+++ tests/custom_test_template.php
@@ -3,4 +3,4 @@
 /**
  * This file is used to test cutom templated objects.
  */
-class custom_test_template extends \Clarkson_Core\Object\Clarkson_Object{}; //phpcs:ignore
+class custom_test_template extends \Clarkson_Core\WordPress_Object\Clarkson_Object{}; //phpcs:ignore
